requirements.txt
# Core TDLib Python bindings
python-telegram==0.15.0
pytdlib==0.1.8

# Async support
asyncio==3.4.3
aiofiles==23.2.1
aiohttp==3.9.1

# Audio processing
pyaudio==0.2.14
numpy==1.26.4
#scipy==1.11.4
soundfile==0.12.1
pydub==0.25.1
webrtcvad==2.0.10

# Speech recognition & synthesis
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.0
SpeechRecognition==3.10.0
gtts==2.4.0

# AI/LLM
openai==1.3.0
anthropic==0.7.0
transformers==4.35.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
